change to redux-persist for state management

+++ app/src/tests/state/onAppStart/IAPPLayersNoRacesInError.test.ts
@@ -18,7 +18,7 @@ describe('Can load IAPP layer regardless of ID call or GeoJSON call happening fi
   const overRideHandlers = [
     {
       method: 'get',
-      url: process.env['IAPP_GEOJSON_URL'],      
+      url: process.env['IAPP_GEOJSON_URL'],
       req: null,
       responseBody: IAPPS3Repsonse_Mock(null),
       status: 200
@@ -28,7 +28,9 @@ describe('Can load IAPP layer regardless of ID call or GeoJSON call happening fi
 
   // There might be a better way but this seems to work ok:
   beforeAll(async () => {
-    localStorage.clear();
+    if (localStorage) {
+      localStorage.clear();
+    }
     require('../../../main');
     await waitFor(() => {
       expect(store).toBeDefined();
